pyproject.toml
[tool.poetry]
name = "style-seeker-ai"
version = "0.1.0"
description = ""
authors = ["Harold Arpanet"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.5"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
markdown = "^3.7"
prefect = "^3.1.13"
meilisearch = "^0.33.1"
minio = "^7.2.15"
python-decouple = "^3.8"
openai = "^1.59.9"
python-telegram-bot = "^21.10"
django-cors-headers = "^4.6.0"
gunicorn = "^23.0.0"
whitenoise = "^6.8.2"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
